pyproject.toml
[project]
name = "doc-improver"
version = "0.1.0"
description = "A document improvement tool"
authors = [
    {name = "Dr Ash Booth"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-docx (>=1.1.2,<2.0.0)",
    "python-multipart (>=0.0.9)",
    "openai (>=1.74.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "Pillow (>=10.2.0,<11.0.0)",  # For image handling in tests
    "pypandoc (>=1.13,<2.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "lxml (>=5.2.2,<6.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
